refactor: expose testutils from ITS crate
