+++ contracts/interchain-token-service/tests/deploy_remote_canonical_token.rs
@@ -1,5 +1,3 @@
-mod utils;
-
 use soroban_sdk::testutils::{Address as _, AuthorizedFunction, AuthorizedInvocation};
 use soroban_sdk::token::{self, StellarAssetClient};
 use soroban_sdk::{Address, Bytes, BytesN, IntoVal, String, Symbol};
@@ -7,10 +5,10 @@ use soroban_token_sdk::metadata::TokenMetadata;
 use stellar_axelar_std::address::AddressExt;
 use stellar_axelar_std::{auth_invocation, events, mock_auth};
 use stellar_interchain_token_service::event::InterchainTokenDeploymentStartedEvent;
+use stellar_interchain_token_service::testutils::{setup_env, setup_gas_token, TokenMetadataExt};
 use stellar_interchain_token_service::types::{
     DeployInterchainToken, HubMessage, Message, TokenManagerType,
 };
-use utils::{setup_env, setup_gas_token, TokenMetadataExt};
 
 #[test]
 fn deploy_remote_canonical_token_succeeds() {
