+++ contracts/interchain-token-service/tests/executable.rs
@@ -1,14 +1,12 @@
-mod utils;
-
 use soroban_sdk::testutils::Address as _;
-use soroban_sdk::{token, vec, Address, Bytes, BytesN, String};
+use soroban_sdk::{token, vec, Address, Bytes, BytesN, IntoVal, String};
 use stellar_axelar_gateway::testutils::{generate_proof, get_approve_hash};
 use stellar_axelar_gateway::types::Message as GatewayMessage;
 use stellar_axelar_std::address::AddressExt;
 use stellar_axelar_std::traits::BytesExt;
 use stellar_axelar_std::{assert_auth_err, events};
+use stellar_interchain_token_service::testutils::{setup_env, setup_its_token};
 use stellar_interchain_token_service::types::{HubMessage, InterchainTransfer, Message};
-use utils::{setup_env, setup_its_token};
 
 mod test {
     use core::fmt::Debug;
