+++ contracts/interchain-token-service/tests/flow_limit.rs
@@ -1,7 +1,5 @@
-mod utils;
-
 use soroban_sdk::testutils::{Address as _, Ledger as _};
-use soroban_sdk::{vec, Address, Bytes, BytesN, Env, String};
+use soroban_sdk::{vec, Address, Bytes, BytesN, Env, IntoVal, String};
 use stellar_axelar_gateway::testutils::TestSignerSet;
 use stellar_axelar_gateway::types::Message as GatewayMessage;
 use stellar_axelar_gateway::AxelarGatewayClient;
@@ -10,11 +8,11 @@ use stellar_axelar_std::traits::BytesExt;
 use stellar_axelar_std::{assert_auth, assert_contract_err, events};
 use stellar_interchain_token_service::error::ContractError;
 use stellar_interchain_token_service::event::FlowLimitSetEvent;
-use stellar_interchain_token_service::types::{HubMessage, InterchainTransfer, Message};
-use stellar_interchain_token_service::InterchainTokenServiceClient;
-use utils::{
+use stellar_interchain_token_service::testutils::{
     approve_gateway_messages, register_chains, setup_env, setup_gas_token, setup_its_token,
 };
+use stellar_interchain_token_service::types::{HubMessage, InterchainTransfer, Message};
+use stellar_interchain_token_service::InterchainTokenServiceClient;
 
 struct GatewayConfig<'a> {
     client: AxelarGatewayClient<'a>,
