+++ contracts/interchain-token-service/tests/pause.rs
@@ -1,10 +1,7 @@
-mod utils;
-
 use soroban_sdk::testutils::Address as _;
-use soroban_sdk::Address;
+use soroban_sdk::{Address, IntoVal};
 use stellar_axelar_std::{assert_auth, assert_auth_err, events};
-use stellar_interchain_token_service::event::PauseStatusSetEvent;
-use utils::setup_env;
+use stellar_interchain_token_service::{event::PauseStatusSetEvent, testutils::setup_env};
 
 #[test]
 fn pause_succeeds() {
